feat(spy): Offer UX redesign + intelligence fields + session docs
Pending changes from prior sessions (Vision-1):
- OfertaCard: intelligence badges (funnel_type, creative_angle, relevance_score)
- SpyOverviewTab: screenshot field, intelligence card, scale signals
- SpyOfferDetail: 7-tab layout with DataMetricCard grid
- DataMetricCard: design system refinements
- AppHeader: layout adjustment
- types.ts: regenerated from Supabase schema
- Materialized views migration: minor fix
- Session docs from 2026-03-04 and 2026-03-10

+++ supabase/migrations/20260220220000_add_materialized_views.sql
@@ -23,7 +23,7 @@ CREATE MATERIALIZED VIEW mv_offer_traffic_summary AS
 SELECT
   spied_offer_id,
   COUNT(DISTINCT domain)                                        AS domain_count,
-  SUM(visits)                                                   AS total_visits,
+  SUM(visits::BIGINT)                                           AS total_visits,
   MAX(period_date)                                              AS latest_period,
   MIN(period_date)                                              AS earliest_period,
   MAX(CASE WHEN period_type = 'monthly_sw' THEN visits END)    AS latest_sw_visits,
